refactor(eval): 拆分 harness 过 sentrux 棘轮（functions>25 巨石规则）(#93)
Refs #93
