+++ eval/reporting.py
@@ -1,0 +1,143 @@
+#!/usr/bin/env python3
+"""Aggregation and Markdown rendering for eval/harness.py's baseline output.
+
+Split out of eval/harness.py (issue #93) so no single file in eval/ trips the
+sentrux god-file ratchet (`functions > 25 && loc > 400`, or `loc > 800`;
+crates/code-intel-cli/src/sentrux_gate.rs:738). Pure functions over
+already-computed result/baseline dicts -- no file I/O, no mutable module
+state, no dependency on eval/harness.py or eval/arms.py.
+"""
+
+from __future__ import annotations
+
+from typing import Any, Dict, List, Sequence
+
+
+def aggregate(results: Sequence[Dict[str, Any]]) -> Dict[str, Any]:
+    total = len(results)
+
+    def arm_summary(key: str) -> Dict[str, Any]:
+        covered = [r for r in results if r[key]["covered"]]
+        total_bytes = sum(r[key]["bytes_to_answer"] for r in results)
+        covered_bytes = sum(r[key]["bytes_to_answer"] for r in covered)
+        return {
+            "covered_count": len(covered),
+            "coverage_rate": round(len(covered) / total, 4) if total else 0.0,
+            "total_bytes_all_questions": total_bytes,
+            "total_bytes_when_covered": covered_bytes,
+            "mean_bytes_when_covered": (
+                round(covered_bytes / len(covered), 2) if covered else None
+            ),
+        }
+
+    categories = sorted({r["category"] for r in results})
+    by_category = {}
+    for cat in categories:
+        rows = [r for r in results if r["category"] == cat]
+        by_category[cat] = {
+            "count": len(rows),
+            "arm_a_covered": sum(1 for r in rows if r["arm_a"]["covered"]),
+            "arm_b_covered": sum(1 for r in rows if r["arm_b"]["covered"]),
+        }
+
+    reasons_a: Dict[str, int] = {}
+    for r in results:
+        reason = r["arm_a"]["reason"]
+        if reason:
+            reasons_a[reason] = reasons_a.get(reason, 0) + 1
+    reasons_b: Dict[str, int] = {}
+    for r in results:
+        reason = r["arm_b"]["reason"]
+        if reason:
+            reasons_b[reason] = reasons_b.get(reason, 0) + 1
+
+    win_loss = {
+        "a_wins": sum(1 for r in results if r["winner"] == "A"),
+        "b_wins": sum(1 for r in results if r["winner"] == "B"),
+        "ties": sum(1 for r in results if r["winner"] == "tie"),
+        "neither_covered": sum(1 for r in results if r["winner"] == "neither"),
+    }
+
+    return {
+        "total_questions": total,
+        "arm_a": arm_summary("arm_a"),
+        "arm_b": arm_summary("arm_b"),
+        "by_category": by_category,
+        "arm_a_failure_reasons": reasons_a,
+        "arm_b_failure_reasons": reasons_b,
+        "win_loss": win_loss,
+        "win_loss_table": [
+            {
+                "id": r["id"],
+                "category": r["category"],
+                "arm_a_covered": r["arm_a"]["covered"],
+                "arm_a_bytes": r["arm_a"]["bytes_to_answer"],
+                "arm_b_covered": r["arm_b"]["covered"],
+                "arm_b_bytes": r["arm_b"]["bytes_to_answer"],
+                "winner": r["winner"],
+            }
+            for r in results
+        ],
+    }
+
+
+def render_markdown(baseline: Dict[str, Any]) -> str:
+    agg = baseline["aggregate"]
+    a, b = agg["arm_a"], agg["arm_b"]
+    total = agg["total_questions"]
+
+    lines: List[str] = []
+    lines.append("# Eval v1 baseline -- artifact-guided (A) vs naive (B)")
+    lines.append("")
+    lines.append(
+        f"Generated by `python eval/harness.py --render` from "
+        f"`{baseline['meta']['baseline_file']}`. Do not hand-edit; re-run harness.py instead."
+    )
+    lines.append("")
+    lines.append(f"HEAD: `{baseline['meta']['repo_head_sha']}` | questions: {total}")
+    lines.append("")
+    lines.append("| id | category | A covered | A bytes | B covered | B bytes | winner |")
+    lines.append("|---|---|---|---|---|---|---|")
+    for row in agg["win_loss_table"]:
+        lines.append(
+            f"| {row['id']} | {row['category']} | "
+            f"{'yes' if row['arm_a_covered'] else 'no'} | {row['arm_a_bytes']:,} | "
+            f"{'yes' if row['arm_b_covered'] else 'no'} | {row['arm_b_bytes']:,} | "
+            f"{row['winner']} |"
+        )
+    lines.append("")
+    lines.append("## Findings")
+    lines.append("")
+
+    wl = agg["win_loss"]
+    a_reasons = ", ".join(f"{k}={v}" for k, v in sorted(agg["arm_a_failure_reasons"].items())) or "none"
+    b_reasons = ", ".join(f"{k}={v}" for k, v in sorted(agg["arm_b_failure_reasons"].items())) or "none"
+    cat_bits = "; ".join(
+        f"{cat}: A {row['arm_a_covered']}/{row['count']}, B {row['arm_b_covered']}/{row['count']}"
+        for cat, row in sorted(agg["by_category"].items())
+    )
+
+    lines.append(
+        f"1. Arm A covered {a['covered_count']}/{total} questions "
+        f"({a['total_bytes_when_covered']:,} bytes over those, "
+        f"{a['total_bytes_all_questions']:,} bytes total including failed searches); "
+        f"Arm B covered {b['covered_count']}/{total} "
+        f"({b['total_bytes_when_covered']:,} bytes over those, "
+        f"{b['total_bytes_all_questions']:,} bytes total)."
+    )
+    lines.append(
+        f"2. Head-to-head: A wins {wl['a_wins']}, B wins {wl['b_wins']}, "
+        f"ties {wl['ties']}, neither covered {wl['neither_covered']} (out of {total})."
+    )
+    lines.append(f"3. Arm A failure reasons: {a_reasons}. Arm B failure reasons: {b_reasons}.")
+    lines.append(f"4. By category (covered/total): {cat_bits}.")
+    setup = baseline["setup"]
+    meta_timing = baseline["meta"]["timing"]
+    lines.append(
+        f"5. One-time setup: `cargo build --release` + `code-intel . --mode lite` produced a "
+        f"{setup['setup_bytes']:,}-byte artifact corpus in "
+        f"{meta_timing['build_seconds'] + meta_timing['pipeline_run_seconds']:.1f}s "
+        "(amortized across all 12 questions, not charged per-question)."
+    )
+    lines.append("")
+    return "\n".join(lines) + "\n"
